[refactoring] split Exit.h & SsdMock.h
