+++ Shell/Exit.h
@@ -1,0 +1,19 @@
+#pragma once
+#include <cstdlib>
+
+class iExit
+{
+public:
+    virtual void doExit() = 0;
+private:
+};
+
+class Exit : public iExit
+{
+public:
+    void doExit() override
+    {
+        exit(0);
+    }
+private:
+};
